Better parsing

+++ bird-protocol-versions/src/lib.rs
@@ -1,13 +1,15 @@
 use std::ops::Range;
-use bird_protocol::{ProtocolReadable, ProtocolSize, ProtocolWritable};
+use bird_protocol::{ProtocolReadable, ProtocolSize, ProtocolWritable, VarInt};
 use bird_protocol::derive::{ProtocolReadable, ProtocolWritable};
 
 #[derive(ProtocolWritable, ProtocolReadable)]
 pub struct Handshake<'a> {
+    #[bp(variant = VarInt)]
     pub protocol_version: i32,
     pub server_address: &'a str,
     pub server_port: u16,
-    pub next_state: u8,
+    #[bp(variant = VarInt)]
+    pub next_state: i32,
 }
 
 impl ProtocolSize for Handshake<'_> {
